Module my_project::string_task::frequency

frequency module includes function that takes strings and calculate character frequency and perform task accordingly

Functions

calculate_frequency

function to calculate frequency of characters

frequency_main

Function to read two strings from user, separate characters based on frequency and replace underscore(_) in replace_main function